made wwpdb entries more accessible
