+++ prody/proteins/wwpdb.py
@@ -8,9 +8,9 @@ from os.path import isdir, isfile, join, split, splitext, normpath
 
 from prody import LOGGER, SETTINGS
 from prody.utilities import makePath, gunzip, relpath, copyFile, openURL
-from prody.utilities import sympath
+from prody.utilities import sympath, checkIdentifiers
 
-__all__ = ['wwPDBServer', 'fetchPDBviaFTP', 'fetchPDBviaHTTP']
+__all__ = ['wwPDBServer', 'fetchPDBviaFTP', 'fetchPDBviaHTTP', 'WWPDB_FTP_SERVERS']
 
 
 _WWPDB_RCSB = ('RCSB PDB (USA)', 'ftp.wwpdb.org', '/pub/')
@@ -86,27 +86,6 @@ def wwPDBServer(*key):
         raise TypeError('one wwPDB server identifier is expected, {0} given'
                         .format(len(key)))
 
-
-def checkIdentifiers(*pdb):
-    """Check whether *pdb* identifiers are valid, and replace invalid ones
-    with **None** in place."""
-
-    identifiers = []
-    append = identifiers.append
-    for pid in pdb:
-        try:
-            pid = pid.strip().lower()
-        except AttributeError:
-            LOGGER.warn('{0} is not a valid identifier.'.format(repr(pid)))
-            append(None)
-        else:
-            if not (len(pid) == 4 and pid.isalnum()):
-                LOGGER.warn('{0} is not a valid identifier.'
-                            .format(repr(pid)))
-                append(None)
-            else:
-                append(pid)
-    return identifiers
 
 
 def fetchPDBviaFTP(*pdb, **kwargs):
